requirements.txt
requests>=2.28.0
Pillow>=10.0.0
pystray>=0.19.5.0
PyYAML>=6.0
pyinstaller
pyzbar